Refactors project structure for clarity
Restructures the project to improve organization and maintainability.

Moves code related to each day's challenge into its own top-level directory.
This simplifies navigation and makes it easier to find the code for
a specific day.

Removes the intermediate "2025/1/Day1" and "2025/2" folder levels.
Updates the solution file to reflect the new project structure.