Adds solutions for Day 6 parts 1 and 2
Implements solutions for Day 6 of the 2015 Advent of Code challenge.

Part 1 calculates the number of lights lit after following a series of instructions to turn on, turn off, or toggle lights in a grid.

Part 2 calculates the total brightness of the lights, where 'turn on' increases brightness by 1, 'turn off' decreases brightness by 1 (minimum 0), and 'toggle' increases brightness by 2.